Fix quit, kill, nick to only send relevant channels; test quit, nick
Previously it wasn't tracking if the users were in the
relevant channels before updating things and then sending
the event. This is fixed in this commit.

Also add tests that various joins, names, parts, quits and
nicks cause it to send the right channels and names with
the events. Not tested is kill.