Fix the bot to send only when the connection is available
This prevents a crash created by the following:

```js
var client = new irc.Client(…);
client.connect();
client.disconnect();
client.end();
client.say('#channel', 'message');
```

This happens if the connection is broken by, say, an
internet connectivity problem (which can trigger the
`end` method of the irc client), but then a message is
attemptedly transmitted.

There is also a new test that breaks before this commit,
ensuring it does not crash in a situation like this.