Extract error and debug to an 'out' object to reduce repeated guarding
There were a lot of instances of:

```javascript
if (self.opt.debug) {
    util.log("debug msg here");
}
```

So this commit aims to reduce that by extracting it to an `out` object
in the bot (`out.debug`, `out.error`) which does the guard for you. In
the case of `out.error`, it also wraps the arguments in the relevant
color coding as often occurred through the code.

This also has `debug` override `showErrors` if true – that is, if
`debug` is true and `showErrors` is false, it will pretend `showErrors`
is true.

`util.log` is apparently deprecated so we should aim to move away from
that in the future also.