Fix issue #19 by adding support for error 477
Sadly, according to
http://defs.ircdocs.horse/defs/numerics.html#err-nochanmodes-477
it looks like error 477 has conflicting usage:

- ERR_NOCHANMODES (by RFC2812 spec)
- ERR_NEEDREGGEDNICK (with Bahamut, ircu, Unreal)

This commit allows codes to just specify a type, instead
of also a name, and adds a test to ensure this code is
parsed as expected.

It also fixes up the order of actual/expected in
test-parse-line.js's t.equal calls, and makes the messages
more grammatical in context (describe the test, not the
success).