Fix .join to use a comma-separated list per IRC spec
As per https://tools.ietf.org/html/rfc2812#page-16, the client is
permitted to use a comma-separated list to communicate the channels it
wishes to join to the server. This is as opposed to a space-separated
list, as stated in the docs, which fails to do anything but join the
first channel on the servers I tested it on.

The server is *not* permitted to do the same back to the client (though
it only "should not", as opposed to being explicitly forbidden from
doing so.).

This commit also adds tests for the `.join` method of the `irc.Client`
API.