Add 378 RPL_WHOISHOST to the codes list
This fixes part of #38.

It appears to have a conflicting definition (see
https://www.alien.net.au/irc/irc2numerics.html), but this is (1) with a
deprecated AustHex response, or (2) a response from aircd. Seeing as I
cannot *find* aircd, and it does not appear when searching around for
"aircd irc" on DuckDuckGo, I'm proceeding in adding this code as an
unambiguous Unreal code.

If a conflict is found at some future point, I can look into adding to
`parse_message` to check heuristically which of a set of options an
ambiguous code might be. (For example, 378 in Unreal seems to always
start its third argument with "is connecting from".)